Extract method extractEvent from Class Event to Class Parser
